Move duration formatter and tictoc to datetime
